dev: move gaussian_setup fixture from sbc only to global fixtures
